+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.6.2] - 2024-06-06
+#### Added
+- Cleaner OpenAI error handling for common issues
+
+### [0.6.1] - 2024-06-06
+#### Removed
+- Response log for image generation
+
 ### [0.6.0] - 2024-06-06
 #### Added
 - Ability to override the OpenAI endpoint with OPENAI_API_BASE_URL env var